Fixed bug with Long mapping in ID

+++ src/main/java/com/montini/teamsports/dao/LocationDao.java
@@ -5,7 +5,7 @@ import com.montini.teamsports.model.Location;
 import java.util.List;
 
 public interface LocationDao {
-    public Location getLocation(long id);
+    public Location getLocation(Integer id);
     public List<Location> getAllLocation();
     public void deleteLocation(Location id);
     public void updateLocation(Location location);
